backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.2
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
polygon-api-client==1.12.3
pandas==2.1.3
numpy==1.26.2
ta==0.10.2  # Technical analysis library
httpx==0.25.2
alembic==1.12.1
pytest==7.4.3
pytest-asyncio==0.21.1 